requirements.txt
absl-py==2.2.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
boto3==1.37.20
botocore==1.37.20
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.3.2
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
durationpy==0.9
et_xmlfile==2.0.0
executing==2.2.0
fake-useragent==2.1.0
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
google-auth==2.38.0
google-genai==1.5.0
google-pasta==0.2.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.0
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.21
langchain-chroma==0.2.2
langchain-community==0.3.20
langchain-core==0.3.48
langchain-openai==0.3.10
langchain-text-splitters==0.3.7
langsmith==0.3.18
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.31.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.68.2
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
opt_einsum==3.4.0
optree==0.14.1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.1
posthog==3.21.0
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.0b1
pydantic-settings==2.8.1
pydantic_core==2.31.1
Pygments==2.19.1
pyparsing==3.2.1
pypdf==5.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
squarify==0.4.4
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vnai==2.0.1
vnstock==3.2.1
vnstock3==3.1.0
vnstock_ezchart==0.0.2
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wordcloud==1.9.4
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
